+++ common/sys/src/state.rs
@@ -208,8 +208,10 @@ impl State {
         #[cfg(feature = "plugins")]
         ecs.insert(match PluginMgr::from_assets() {
             Ok(plugin_mgr) => {
-                if let Err(e) = plugin_mgr
-                    .execute_event("on_load", &plugin_api::event::PluginLoadEvent { game_mode })
+                if let Err(e) =
+                    plugin_mgr.execute_event(&ecs, "on_load", &plugin_api::event::PluginLoadEvent {
+                        game_mode,
+                    })
                 {
                     tracing::error!(?e, "Failed to run plugin init");
                     tracing::info!(